requirements.txt
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
gcloud==0.18.3
google-api-core==1.16.0
google-api-python-client==1.7.11
google-auth==1.11.2
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
importlib-metadata==1.5.0
more-itertools==8.2.0
oauth2client==4.1.3
packaging==20.3
pluggy==0.13.1
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
pytest==5.3.5
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
zipp==3.1.0
